# libplugin-pay: use map for channel hints (on top of #7494)