--- a/common/route.c
+++ b/common/route.c
@@ -94,10 +94,6 @@ static bool dijkstra_to_hops(struct route_hop **hops,
 	/* OK, populate other fields. */
 	c = dijkstra_best_chan(dij, curidx);
 
-	gossmap_chan_get_capacity(gossmap, c, &total);
-	if (!amount_sat_to_msat(&(*hops)[num_hops].total_amount, total))
-		abort();
-
 	if (c->half[0].nodeidx == curidx) {
 		(*hops)[num_hops].direction = 0;
 	} else {
@@ -113,6 +109,8 @@ static bool dijkstra_to_hops(struct route_hop **hops,
 	if (!dijkstra_to_hops(hops, gossmap, dij, next, amount, cltv))
 		return false;
 
+	gossmap_chan_get_capacity(gossmap, c, &total);
+	(*hops)[num_hops].total_amount.millisatoshis = total.satoshis * 1000; /* Raw: simpler. */
 	(*hops)[num_hops].amount = *amount;
 	(*hops)[num_hops].delay = *cltv;