--- a/common/route.c
+++ b/common/route.c
@@ -105,8 +105,7 @@ static bool dijkstra_to_hops(struct route_hop **hops,
 		return false;
 
 	gossmap_chan_get_capacity(gossmap, c, &total);
-	if (!amount_sat_to_msat(&((*hops[num_hops]).total_amount), total))
-		abort();
+	(*hops)[num_hops].capacity = total;
 	(*hops)[num_hops].amount = *amount;
 	(*hops)[num_hops].delay = *cltv;