--- a/common/route.h
+++ b/common/route.h
@@ -18,6 +18,7 @@ struct gossmap_node;
  * @direction: 0 (dest node_id < src node_id), 1 (dest node_id > src).
  * @node_id: the node_id of the destination of this hop.
  * @amount: amount to send through this hop.
+ * @total_amount: The amount the channel was funded with.
  * @delay: total cltv delay at this hop.
  */
 struct route_hop {
@@ -25,6 +26,7 @@ struct route_hop {
 	int direction;
 	struct node_id node_id;
 	struct amount_msat amount;
+	struct amount_msat total_amount;
 	u32 delay;
 };