--- a/plugins/channel_hint.c
+++ b/plugins/channel_hint.c
@@ -71,6 +71,76 @@ bool channel_hint_update(const struct timeabs now, struct channel_hint *hint)
 	       amount_msat_greater(capacity, hint->estimated_capacity);
 }
 
+struct channel_hint *channel_hint_set_find(struct channel_hint_set *self,
+					   const struct short_channel_id_dir *scidd)
+{
+  for (size_t i=0; i<tal_count(self->hints); i++) {
+    struct channel_hint *hint = &self->hints[i];
+    if (short_channel_id_dir_eq(&hint->scid, scidd))
+      return hint;
+  }
+  return NULL;
+}
+
+/* See header */
+struct channel_hint *
+channel_hint_set_add(struct channel_hint_set *self, u32 timestamp,
+		     const struct short_channel_id_dir *scidd, bool enabled,
+		     const struct amount_msat *estimated_capacity,
+		     const struct amount_sat capacity, u16 *htlc_budget)
+{
+	bool modified = false;
+	struct channel_hint *old, *newhint = tal(tmpctx, struct channel_hint);
+	struct timeabs now = time_now();
+
+	/* If the channel is marked as enabled it must have an estimate. */
+	assert(!enabled || estimated_capacity != NULL);
+	newhint->enabled = enabled;
+	newhint->scid = *scidd;
+	newhint->capacity = capacity;
+	if (estimated_capacity != NULL)
+		newhint->estimated_capacity = *estimated_capacity;
+	newhint->local = NULL;
+	newhint->timestamp = timestamp;
+
+	/* Project the channel_hints into the same domain, so we can merge them.
+	 */
+	channel_hint_update(now, newhint);
+	channel_hint_set_update(self, now);
+
+	/* And now we can merge the new hint into the existing ones if there
+	   are any. */
+	old = channel_hint_set_find(self, scidd);
+	if (old == NULL) {
+		tal_arr_expand(&self->hints, *newhint);
+		// TODO extend the array
+		return &self->hints[tal_count(self->hints) - 1];
+	} else {
+		/* Prefer to disable a channel. */
+		if (!enabled && old->enabled) {
+			old->enabled = false;
+			modified = true;
+		}
+		/* Prefer the more conservative estimate. */
+		if (estimated_capacity != NULL &&
+		    amount_msat_greater(old->estimated_capacity,
+					newhint->estimated_capacity)) {
+			old->estimated_capacity = newhint->estimated_capacity;
+			modified = true;
+		}
+		if (newhint->local) {
+			tal_free(old->local);
+			old->local = tal_steal(old, newhint->local);
+		}
+	}
+	if (modified) {
+		old->timestamp = timestamp;
+		return old;
+	} else {
+		return NULL;
+	}
+}
+
 /**
  * Load a channel_hint from its JSON representation.
  *