… common/test/Makefile
… common/test/run-route-specific.c
… common/test/run-route.c
… plugins/keysend.c
… plugins/libplugin-pay.c
… plugins/libplugin-pay.h
… plugins/pay.c
… plugins/test/Makefile
… plugins/test/run-route-calc.c
… plugins/test/run-route-overlong.c
… tests/test_connection.py
… tests/test_pay.py
… wallet/test/run-wallet.c